定时发送短信

+++ src/main/java/com/tengyue360/dao/SsAccountNumberMapper.java
@@ -37,6 +37,6 @@ public interface SsAccountNumberMapper {
      * @throws Exception
      */
 
-    Long countParent(@Param("role") String role);
+    Integer countParent(@Param("role") String role);
 
 }
